refactor(cda-storage): Split recovery tests into separate module.
